+++ src/main/java/de/umass/lastfm/Tag.java
@@ -50,10 +50,10 @@ public class Tag implements Comparable<Tag> {
 
 	private String name;
 	private String url;
-	private int count;
+	private long count;
 
 	private boolean streamable;
-	private int reach;
+	private long reach;
 
 	private Date wikiLastChanged;
 	private String wikiSummary;
@@ -63,7 +63,7 @@ public class Tag implements Comparable<Tag> {
 		this.name = name;
 	}
 
-	public int getCount() {
+	public long getCount() {
 		return count;
 	}
 
@@ -73,7 +73,7 @@ public class Tag implements Comparable<Tag> {
 	 * @return Number of Taggings
 	 * @see Tag#getInfo(String, String)
 	 */
-	public int getTaggings() {
+	public long getTaggings() {
 		return count;
 	}
 
@@ -89,7 +89,7 @@ public class Tag implements Comparable<Tag> {
 		return streamable;
 	}
 
-	public int getReach() {
+	public long getReach() {
 		return reach;
 	}
 
@@ -226,12 +226,12 @@ public class Tag implements Comparable<Tag> {
 			t.url = element.getChildText("url");
 
 			if (element.hasChild("count"))
-				t.count = Integer.parseInt(element.getChildText("count"));
+				t.count = Long.parseLong(element.getChildText("count"));
 			else if (element.hasChild("taggings"))
-				t.count = Integer.parseInt(element.getChildText("taggings"));
+				t.count = Long.parseLong(element.getChildText("taggings"));
 
 			if (element.hasChild("reach"))
-				t.reach = Integer.parseInt(element.getChildText("reach"));
+				t.reach = Long.parseLong(element.getChildText("reach"));
 			if (element.hasChild("streamable"))
 				t.streamable = StringUtilities.convertToBoolean(element.getChildText("streamable"));
